requirements.txt
adafruit_ssd1306==1.6.2
board==1.0
Pillow==7.1.2
